Lift common document endpoint testing resources to utils
